fix(BCE): R6-R8 代码审计根治 — BCE-038~041 全类根除 + 残留清零
BCE-038: mutex poison 静默吞没 → log::error! + 防御性返回
  - three_tier_swap.rs: migrate_expert_pages/run_defrag_cycle 3处

BCE-039: 路由解析失败返回 false(本地页) → true(需跨节点传输)
  - coordinator/kv.rs: needs_cross_node_transfer 保守安全默认

BCE-040: clamp 上界无效 (usize::MAX as f32 = +inf)
  - scheduler/policy.rs: force_swap_out_count 上界 1000.0

BCE-041: mutex poison 全项目横扫 — .lock().unwrap() → .expect()/log::error!
  - executor.rs, executor_builder.rs, mega_kernel_callback.rs
  - executor_core.inc.rs: 5处 weight_page_table/fault_handler
  - weight_tier.rs: 8处 allocations
  - three_tier_swap.rs: 2处 RwLock poison
  - 残留清零: 生产代码零 .lock().unwrap()，仅测试代码保留

C-5.4 修正: BUG-KNOWLEDGE.md '优雅降级' → '防御性返回'

+++ src/engine/callbacks/moe_dispatch.rs
@@ -2513,9 +2513,9 @@ mod tests {
             destination: ExpertWeightLocation::GpuVram,
             bytes: 0,
             estimated_latency_us: f32::MAX,
-            priority: u32::MAX,
+            priority: u64::MAX,
         };
-        assert_eq!(req.priority, u32::MAX);
+        assert_eq!(req.priority, u64::MAX);
     }
 
     // ── ExpertPrefetchRequest: Debug format ──────────────────────────────
@@ -4870,11 +4870,11 @@ mod tests {
             destination: ExpertWeightLocation::GpuVram,
             bytes: usize::MAX,
             estimated_latency_us: f32::MAX,
-            priority: u32::MAX,
+            priority: u64::MAX,
         };
         assert_eq!(req.expert_idx, usize::MAX);
         assert_eq!(req.bytes, usize::MAX);
-        assert_eq!(req.priority, u32::MAX);
+        assert_eq!(req.priority, u64::MAX);
     }
 
     #[test]
